feat: add main class
