+++ Mammal.java
@@ -1,31 +1,9 @@
-import java.util.Scanner;
-public abstract class Mammal {  //Nombre de la clase Padre
 
-    //Atributos de la clase Padre
+public abstract class Mammal {
+
     public String species;
     public String feeding;
     public String name;
     public String breed;
 
-
-    //Constructor clase Padre
-  //  public Mammal(String NewName){
-   // }
-    //Metodo de la clase Padre
-    public static void main(String[] args) {
-
-        Scanner orderDog = new Scanner(System.in);
-        System.out.println("¿Cuál es el nombre del perro al que busca?");
-        String client = orderDog.next();
-
-        //Funcionaria como una base de datos, quisiera ponerle un if de que si no se encuentra le aparezca otro cartel
-        
-    Canine Perro1 = new Canine("Canine","Omnivore","Lolo", "Dalmata");
-    Canine Perro2 = new Canine("Canine","Omnivore","Pepe", "Pastor Aleman");
-
-
-
-    System.out.println("Excelente, " + client + " será entregado en 5 minutos");
-
-            }
-        }
+}
